+++ mmdet/ops/affine_grid/__init__.py
@@ -1,0 +1,3 @@
+from .affine_grid import affine_grid
+
+__all__ = ['affine_grid']
