--- mmdet/ops/affine_grid_generator/__init__.py
@@ -1,3 +1,0 @@
-from .affine_grid_generator import affine_grid
-
-__all__ = ['affine_grid']
